+++ algolia/recommend/configuration.go
@@ -1,13 +1,14 @@
-package recommendation
+package recommend
 
 import (
 	"time"
 
-	"github.com/algolia/algoliasearch-client-go/v3/algolia/region"
+	"github.com/algolia/algoliasearch-client-go/v3/algolia/compression"
 	"github.com/algolia/algoliasearch-client-go/v3/algolia/transport"
 )
 
-// Deprecated: use personalization.Configuration instead
+// Configuration contains all the different parameters one can change to
+// instantiate a new client for the Recommend API.
 type Configuration struct {
 	AppID          string
 	APIKey         string
@@ -15,7 +16,7 @@ type Configuration struct {
 	Requester      transport.Requester
 	ReadTimeout    time.Duration
 	WriteTimeout   time.Duration
-	Region         region.Region
 	Headers        map[string]string
 	ExtraUserAgent string
+	Compression    compression.Compression
 }