+++ cts/index/exists_test.go
@@ -3,6 +3,7 @@ package index
 import (
 	"testing"
 
+	"github.com/algolia/algoliasearch-client-go/v3/algolia/search"
 	"github.com/algolia/algoliasearch-client-go/v3/cts"
 	"github.com/stretchr/testify/require"
 )
@@ -23,4 +24,8 @@ func TestExists(t *testing.T) {
 	ok, _ = index.Exists()
 	require.True(t, ok)
 	require.NoError(t, err)
+
+	ok, err = search.NewClient(index.GetAppID(), "invalid").InitIndex(index.GetName()).Exists()
+	require.False(t, ok)
+	require.Error(t, err)
 }