+++ cts/testing_utils.go
@@ -18,9 +18,10 @@ import (
 	"github.com/algolia/algoliasearch-client-go/v3/algolia/analytics"
 	"github.com/algolia/algoliasearch-client-go/v3/algolia/compression"
 	"github.com/algolia/algoliasearch-client-go/v3/algolia/insights"
-	"github.com/algolia/algoliasearch-client-go/v3/algolia/recommendation"
+	"github.com/algolia/algoliasearch-client-go/v3/algolia/personalization"
 	"github.com/algolia/algoliasearch-client-go/v3/algolia/region"
 	"github.com/algolia/algoliasearch-client-go/v3/algolia/search"
+	"github.com/algolia/algoliasearch-client-go/v3/algolia/suggestions"
 	"github.com/algolia/algoliasearch-client-go/v3/algolia/transport"
 )
 
@@ -48,7 +49,7 @@ func cleanIndexName(indexName string) string {
 		"+",
 		"^",
 	} {
-		indexName = strings.Replace(indexName, char, "_", -1)
+		indexName = strings.Replace(indexName, char, "_", -1) //nolint:gocritic
 	}
 	return indexName
 }
@@ -73,8 +74,12 @@ func InitInsightsClient(t *testing.T) *insights.Client {
 	return initInsightsClientWith(t, "ALGOLIA_APPLICATION_ID_1", "ALGOLIA_ADMIN_KEY_1")
 }
 
-func InitRecommendationClient(t *testing.T) *recommendation.Client {
-	return initRecommendationClientWith(t, "ALGOLIA_APPLICATION_ID_1", "ALGOLIA_ADMIN_KEY_1")
+func InitPersonalizationClient(t *testing.T) *personalization.Client {
+	return initPersonalizationClientWith(t, "ALGOLIA_APPLICATION_ID_1", "ALGOLIA_ADMIN_KEY_1")
+}
+
+func InitQuerySuggestionsClient1(t *testing.T) *suggestions.Client {
+	return initQuerySuggestionsClientWith(t, "ALGOLIA_APPLICATION_ID_1", "ALGOLIA_ADMIN_KEY_1")
 }
 
 func initInsightsClientWith(t *testing.T, appIDEnvVar, apiKeyEnvVar string) *insights.Client {
@@ -100,9 +105,15 @@ func initAnalyticsClientWith(t *testing.T, appIDEnvVar, apiKeyEnvVar string) *an
 	return c
 }
 
-func initRecommendationClientWith(t *testing.T, appIDEnvVar, apiKeyEnvVar string) *recommendation.Client {
+func initPersonalizationClientWith(t *testing.T, appIDEnvVar, apiKeyEnvVar string) *personalization.Client {
 	appID, key := GetTestingCredentials(t, appIDEnvVar, apiKeyEnvVar)
-	c := recommendation.NewClient(appID, key, region.US)
+	c := personalization.NewClient(appID, key, region.US)
+	return c
+}
+
+func initQuerySuggestionsClientWith(t *testing.T, appIDEnvVar, apiKeyEnvVar string) *suggestions.Client {
+	appID, key := GetTestingCredentials(t, appIDEnvVar, apiKeyEnvVar)
+	c := suggestions.NewClient(appID, key)
 	return c
 }